Reduce duplication 2: spikesort utils
