+++ src/spyglass/spikesorting/utils.py
@@ -1,4 +1,7 @@
+import warnings
+
 import numpy as np
+import scipy.stats as stats
 import spikeinterface as si
 
 from spyglass.common.common_ephys import Electrode
@@ -128,3 +131,122 @@ def get_group_by_shank(
                 sge_key["electrode_id"] = elect
                 sge_keys.append(sge_key.copy())
             sort_group += 1
+
+
+def _init_artifact_worker(
+    recording,
+    zscore_thresh=None,
+    amplitude_thresh_uV=None,
+    proportion_above_thresh=1.0,
+):
+    recording = (
+        si.load_extractor(recording)
+        if isinstance(recording, dict)
+        else recording
+    )
+    # create a local dict per worker
+    return {
+        "recording": recording,
+        "zscore_thresh": zscore_thresh,
+        "amplitude_thresh_uV": amplitude_thresh_uV,
+        "proportion_above_thresh": proportion_above_thresh,
+    }
+
+
+def _compute_artifact_chunk(segment_index, start_frame, end_frame, worker_ctx):
+    recording = worker_ctx["recording"]
+    zscore_thresh = worker_ctx["zscore_thresh"]
+    amplitude_thresh = worker_ctx["amplitude_thresh"]
+    proportion_above_thresh = worker_ctx["proportion_above_thresh"]
+    # compute the number of electrodes that have to be above threshold
+    nelect_above = np.ceil(
+        proportion_above_thresh * len(recording.get_channel_ids())
+    )
+
+    traces = recording.get_traces(
+        segment_index=segment_index,
+        start_frame=start_frame,
+        end_frame=end_frame,
+    )
+
+    # find the artifact occurrences using one or both thresholds, across channels
+    if (amplitude_thresh is not None) and (zscore_thresh is None):
+        above_a = np.abs(traces) > amplitude_thresh
+        above_thresh = (
+            np.ravel(np.argwhere(np.sum(above_a, axis=1) >= nelect_above))
+            + start_frame
+        )
+    elif (amplitude_thresh is None) and (zscore_thresh is not None):
+        dataz = np.abs(stats.zscore(traces, axis=1))
+        above_z = dataz > zscore_thresh
+        above_thresh = (
+            np.ravel(np.argwhere(np.sum(above_z, axis=1) >= nelect_above))
+            + start_frame
+        )
+    else:
+        above_a = np.abs(traces) > amplitude_thresh
+        dataz = np.abs(stats.zscore(traces, axis=1))
+        above_z = dataz > zscore_thresh
+        above_thresh = (
+            np.ravel(
+                np.argwhere(
+                    np.sum(np.logical_or(above_z, above_a), axis=1)
+                    >= nelect_above
+                )
+            )
+            + start_frame
+        )
+
+    return above_thresh
+
+
+def _check_artifact_thresholds(
+    amplitude_thresh, zscore_thresh, proportion_above_thresh
+):
+    """Alerts user to likely unintended parameters. Not an exhaustive verification.
+
+    Parameters
+    ----------
+    zscore_thresh: float
+    amplitude_thresh: float
+        Measured in microvolts.
+    proportion_above_thresh: float
+
+    Return
+    ------
+    zscore_thresh: float
+    amplitude_thresh: float
+    proportion_above_thresh: float
+
+    Raise
+    ------
+    ValueError: if signal thresholds are negative
+    """
+    # amplitude or zscore thresholds should be negative, as they are applied to
+    # an absolute signal
+
+    signal_thresholds = [
+        t for t in [amplitude_thresh, zscore_thresh] if t is not None
+    ]
+    for t in signal_thresholds:
+        if t < 0:
+            raise ValueError(
+                "Amplitude and Z-Score thresholds must be >= 0, or None"
+            )
+
+    # proportion_above_threshold should be in [0:1] inclusive
+    if proportion_above_thresh < 0:
+        warnings.warn(
+            "Warning: proportion_above_thresh must be a proportion >0 and <=1."
+            " Using proportion_above_thresh = 0.01 instead of "
+            + f"{str(proportion_above_thresh)}"
+        )
+        proportion_above_thresh = 0.01
+    elif proportion_above_thresh > 1:
+        warnings.warn(
+            "Warning: proportion_above_thresh must be a proportion >0 and <=1. "
+            "Using proportion_above_thresh = 1 instead of "
+            + f"{str(proportion_above_thresh)}"
+        )
+        proportion_above_thresh = 1
+    return amplitude_thresh, zscore_thresh, proportion_above_thresh
